Add general CapWidth and CheriBusBytes macros
These replace the CAP* and MEM* macros that hard-code the list of
supported widths. Eventually all users should migrate to defining these
macros and we can delete the backwards-compatibility code. BERI will
continue to need to use CAP/CAP128/CAP64 internally but that can now be
an implementation detail that doesn't leak to the generic TagController.