Avoid write flit interleaving in rare edge-case
If a read request is made in the middle of a write burst, this currently
causes the first part of the write burst to be forwarded downstream
(along with the read request), with the possibility for tag lookup
writes to be issued before the rest of the write burst is received,
causing the write flits to be associated with the wrong transactions.

Hotfix this by stopping read request processing in TagControllerAXI when
in the middle of a write burst, ensuring that all write flits are seen
back-to-back (without any read request flits in between) by the tag
controller itself.