Bag: Only insert once if multiple entries have the same key
On reset, all keys are ?, so even though we never add duplicate keys we
can start off with them. If we insert an item matching one of these
duplicated keys then all the entries are currently set to valid, rather
than just the first. This ultimately causes the levels of mastReqs and
mastReqIds to not always be in sync in CacheCorderer, causing CacheCore
to attempt to enqueue when mastReqs is full.