MultiLevelTagLookup: Fix byte-enables for zeroing tag writes
This regressed in 8c05a3d3, the first commit for supporting capabilities
split across multiple flits. Revert it back to turning bits into bytes,
as it is in cheri-cpu. Otherwise we fail to zero out half of a line (or
one byte less than half a line if byteOffset is for the upper half).