Changes to TagController to work with 32-bit tags and 64-byte lines.
This is is so that we can have 1 captag + 3 version tags for an experimental Toooba with memory versioning.

There were some assumptions about the relative size of types so needed to make some tweaks.

Also the calculation of tag table structure was missing 32-byte alignment requirement compared to python script.