Replace write-though TC with simple writeandset TC
The write-through tag controller didn't behave as expected because
although CheriMemRequest allows bit level enabling of writes, AXI
does not. BUT since writing tags requires setting individual bits,
the rest of the byte needs to be known when crafting the write
request.

Therefore created a much simpler write-and-set tag controller,
which writes data to memory and sets the root & leaf level tags for
that address. It also zeros the entire root level table.

(No need to handle reads as just used for init)